renamed rpm_pkg into rpm -> more relevant
