+++ python/lzreposync/src/lzreposync/rpm.py
